+++ Extensions/RecluseEdit.Extensions.React/Syntaxes/JsxSyntaxDefinition.cs
@@ -53,8 +53,7 @@ public static class JsxSyntaxDefinition
                     <Begin>`</Begin>
                     <End>`</End>
                     <RuleSet>
-                        <Span begin="\\`" end="" />
-                        <Span begin="\\\\" end="" />
+                        <Span begin="\\" end="." />
                         <Span color="Variable" multiline="true">
                             <Begin>\$\{</Begin>
                             <End>\}</End>
@@ -67,8 +66,7 @@ public static class JsxSyntaxDefinition
                     <Begin>&quot;</Begin>
                     <End>&quot;</End>
                     <RuleSet>
-                        <Span begin="\\&quot;" end="" />
-                        <Span begin="\\\\" end="" />
+                        <Span begin="\\" end="." />
                     </RuleSet>
                 </Span>
 
@@ -77,8 +75,7 @@ public static class JsxSyntaxDefinition
                     <Begin>'</Begin>
                     <End>'</End>
                     <RuleSet>
-                        <Span begin="\\'" end="" />
-                        <Span begin="\\\\" end="" />
+                        <Span begin="\\" end="." />
                     </RuleSet>
                 </Span>